backend/app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
sqlalchemy==2.0.30
pydantic==2.7.0
python-dotenv==1.0.1
alembic==1.13.1
pytest==8.3.2
httpx==0.27.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

